+++ resources-library/src/posixMain/kotlin/Resource.kt
@@ -3,12 +3,14 @@ package com.goncalossilva.resources
 import kotlinx.cinterop.ByteVar
 import kotlinx.cinterop.allocArray
 import kotlinx.cinterop.memScoped
+import kotlinx.cinterop.readBytes
 import kotlinx.cinterop.toKString
 import platform.posix.F_OK
 import platform.posix.access
 import platform.posix.fclose
 import platform.posix.fgets
 import platform.posix.fopen
+import platform.posix.fread
 import platform.posix.posix_errno
 import platform.posix.strerror
 
@@ -29,6 +31,22 @@ public actual class Resource actual constructor(private val path: String) {
             fclose(file)
         }
     }
+
+    public actual fun readBytes(): ByteArray = mutableListOf<Byte>().apply {
+        val file = fopen(path, "r")
+            ?: throw FileReadException("$path: Open failed: ${strerror(posix_errno())}")
+        try {
+            memScoped {
+                val buffer = allocArray<ByteVar>(BUFFER_SIZE)
+                do {
+                    val size = fread(buffer, 1, BUFFER_SIZE.toULong(), file)
+                    addAll(buffer.readBytes(size.toInt()).asIterable())
+                } while (size > 0u)
+            }
+        } finally {
+            fclose(file)
+        }
+    }.toByteArray()
 
     private companion object {
         private const val BUFFER_SIZE = 8 * 1024